moved shared fixtures to conftest.py
